+++ src/main/java/com/acme/yupanaapi/domain/repository/FlowRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.acme.yupanaapi.domain.model.Flow;
 
-public interface FlowRepository extends JpaRepository<Flow, Long>{
-	List<Flow> findAllByWalletId(Long id);
-	List<Flow> findAllByWalletIdAndDeadlineDate(Long walletId, Date date);
+public interface FlowRepository extends JpaRepository<Flow,Integer>{
+	List<Flow> findAllByWalletId(Integer id);
+	List<Flow> findAllByWalletIdAndDeadlineDate(Integer walletId, Date date);
 }
