+++ src/main/java/com/acme/yupanaapi/domain/service/SellerService.java
@@ -4,11 +4,11 @@ import com.acme.yupanaapi.domain.model.Seller;
 import org.springframework.http.ResponseEntity;
 
 public interface SellerService {
-    Seller getSellerByIdAndUserId(Long sellerId, Long userId);
-    Seller getSellerById(Long sellerId);
+    Seller getSellerByIdAndUserId(Integer sellerId, Integer userId);
+    Seller getSellerById(Integer sellerId);
     Seller getSellerByBusinessName(String businessName);
-    Seller createSeller(Seller seller, Long userId);
-    Seller updateSeller(Long sellerId, Long userId, Seller sellerRequest);
-    ResponseEntity<?> deleteSeller(Long sellerId);
+    Seller createSeller(Seller seller, Integer userId);
+    Seller updateSeller(Integer sellerId, Integer userId, Seller sellerRequest);
+    ResponseEntity<?> deleteSeller(Integer sellerId);
 }
