+++ src/main/java/com/acme/yupanaapi/domain/service/TransactionService.java
@@ -7,11 +7,11 @@ import org.springframework.http.ResponseEntity;
 import com.acme.yupanaapi.domain.model.Transaction;
 
 public interface TransactionService {
-	Transaction createTransaction(Transaction transactionEntity, Long flowId);
-	Transaction AssignTransactionWithSale(Long transactionId, Long saleId);
-	Transaction updateTransaction(Transaction transactionEntity, Long transactionId);
-	ResponseEntity<?> deleteTransaction(Long transactionId);
-	Transaction getTransactionById(Long transactionId);
-	List<Transaction> getAllByFlowId(Long flowId);
-	List<Transaction> getAllByHistorialId(Long historialId);
+	Transaction createTransaction(Transaction transactionEntity, Integer flowId);
+	Transaction AssignTransactionWithSale(Integer transactionId, Integer saleId);
+	Transaction updateTransaction(Transaction transactionEntity, Integer transactionId);
+	ResponseEntity<?> deleteTransaction(Integer transactionId);
+	Transaction getTransactionById(Integer transactionId);
+	List<Transaction> getAllByFlowId(Integer flowId);
+	List<Transaction> getAllByHistorialId(Integer historialId);
 }
